Move SubscriptionAmendmentAnalyseResult to its own file
