+++ lambda/src/main/scala/pricemigrationengine/model/AmendmentHandlerHelper.scala
@@ -287,29 +287,6 @@ object AmendmentHandlerHelper {
       case SupporterPlus2026N5    => itIsFewDaysAfterNotification(item)
     }
   }
-
-}
-
-sealed trait SubscriptionAmendmentAnalyseResult
-
-// "SAAR" means "Subscription Amendment Analyse Result"
-
-object SAARReadyToAmend extends SubscriptionAmendmentAnalyseResult
-object SAARCancelledInZuora extends SubscriptionAmendmentAnalyseResult
-object SAARExcludeFromMigration extends SubscriptionAmendmentAnalyseResult
-object SAARFailNoisily extends SubscriptionAmendmentAnalyseResult
-
-object SubscriptionAmendmentAnalyseResult {
-
-  def toString(result: SubscriptionAmendmentAnalyseResult): String = {
-    result match {
-      case SAARReadyToAmend         => "SAARReadyToAmend"
-      case SAARCancelledInZuora     => "SAARCancelledInZuora"
-      case SAARExcludeFromMigration => "SAARExcludeFromMigration"
-      case SAARFailNoisily          => "SAARFailNoisily"
-    }
-  }
-
   def subscriptionIsAmendableSupporterPlus2026(
       item: CohortItem,
       subscription: ZuoraSubscription,
